fix(core): 修复 sqlite_db.cpp 中的 bind_blob 函数参数类型
- 将 bind_blob 函数的 size 参数类型从 uint 改为 size_t
- 在 sqlite_db.cpp 文件中添加了 #include <cstdint> 以支持 size_t 类型

+++ src/core/components/sqlite_db.cpp
@@ -2,6 +2,7 @@
 #include <stdexcept>
 #include <string>
 #include <vector>
+#include <cstdint>
 
 #include "gdsqlite/sqlite/sqlite3.h"
 #include "sqlite_db.h"
@@ -128,7 +129,7 @@ void SQLiteDB::Stmt::bind_double(int idx, double val) {
 	sqlite3_bind_double(stmt_, idx, val);
 }
 
-void SQLiteDB::Stmt::bind_blob(int idx, const void *val, uint size) {
+void SQLiteDB::Stmt::bind_blob(int idx, const void *val, size_t size) {
 	if (!stmt_) {
 		return;
 	}
